pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = ["Serchaev <german.serchaevk2.k2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.109.2"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
email-validator = "^2.1.0.post1"
pymongo = "^4.6.2"
motor = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
